+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v3.0.0 11/30/2018 3:58:11 PM
 #### Changed
 - Directory Structure
+- All legacy templates now are building templates with retrofit code
 
 #### Added
 - git related files